Generating dimples via direct vertice modification
Instead of cutting dimples by creating additional spheres, I now generate them directly by adjusting the coordinates of the vertices on the original sphere. This approach greatly speeds up the generation process compared to the previous method, but it makes the result more dependent on the fineness of the original mesh (less fine grid will make the hole look extremely rough), and will also cause the dimple boundaries to be ambiguous.